# python/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Floor Memory",
    "description": "The set APIs are used to develop Floor pds which can be used as their personal assistants. \n\nThis set of APIs are divided into two parts. \n- Memory and \n- Registration.\n \nThe developer has two ways of using the APIs for the app development. Developer can choose to the  Registration APIs for using the existing xfloor infracture or can implement custom Registration process. In the case of custom registration process, the developer is bound to provide proper authentication mechanisms and then send the user information to xlfoor.",
    "version": "1.0.0",
    "summary": "",
    "license": {
      "name": "MIT",
      "url": "https://xfloor.ai/ipomo/blog1551084548304/1642341504453"
    },
    "termsOfService": "https://xfloor.ai/ipomo/blog1551084548304/1642341504453",
    "contact": {
      "name": "Ipomo",
      "email": "contact@ipomo.in",
      "url": "https://xfloor.ai/ipomo/blog1551084548304/"
    }
  },
...
  "paths": {
...
    "/api/memory/events": {
      "post": {
        "summary": "Create Event (Post Content)",
        "deprecated": false,
        "description": "Posts into the given floor_id. This is asynchronous ingestion. 200 OK means queued, not immediately retrievable.\nThis API allows a user to **post personal content into their POD (Personal Object Database)**.\nThe posted content is stored under a specified **floor**, embedded by the platform, and made available for **semantic querying, conversational retrieval, and memory-based interactions**.\n\nIt is primarily used to:\n\n* Save reminders, notes, writeups, or personal knowledge\n* Upload content that the user wants the system to remember\n* Add information that should later be discoverable via conversational queries\n\nThe content may consist of **text only** or **text combined with one or more media files**.\n\n---\n\n## **Key Capabilities**\n\n* Stores user-generated content inside a specific floor\n* Supports **multi-modal inputs** (text + media)\n* Automatically embeds content for semantic search\n* Makes content available to:\n\n  * `/agent/memory/query`\n  * conversational agents\n  * future recall and analytics\n* Associates content with user, block, and application context\n\n---\n\n## **Authentication**\n\n* Requires a valid, authenticated `user_id`\n* The calling application is responsible for user authentication\n* `app_id` identifies the application context\n\n---\n\n## **Request Type**\n\n**Content-Type:** `multipart/form-data`\n\n---\n\n## **Request Parameters**\n\n### **1. Files (Optional)**\n\n| Field   | Type   | Required | Description                                                           |\n| ------- | ------ | -------- | --------------------------------------------------------------------- |\n| `files` | file[] | Optional | Media files to attach to the content. Multiple files may be uploaded. |\n\n**Supported formats include (but are not limited to):**\n\n* Images: `jpg`, `png`\n* Audio: `mp3`\n* Documents: `pdf`\n* Video: `mp4`\n\nThese files are processed and embedded along with the textual content where applicable.\n\n---\n\n### **2. Input Information (Required)**\n\n| Field        | Type          | Required | Description                                                       |\n| ------------ | ------------- | -------- | ----------------------------------------------------------------- |\n| `input_info` | string (JSON) | Yes      | JSON string containing metadata and textual content for the post. |\n\n---\n\n## **`input_info` Structure**\n\n```json\n{\n  \"floor_id\": \"my_floor\",\n  \"BID\": \"17845683456\",\n  \"user_id\": \"145623907625\",\n  \"title\": \"My floor\",\n  \"description\": \"My floor details\",\n  \"app_id\": \"165434879028\"\n}\n```\n\n---\n\n### **Field Descriptions**\n\n| Field         | Type   | Required | Description                                                                        |\n| ------------- | ------ | -------- | ---------------------------------------------------------------------------------- |\n| `floor_id`    | string | Yes      | Identifier of the user’s floor (POD) where the content will be stored.             |\n| `block_type`  | string | Yes      | Type of block under which the content is categorized (e.g., post, note, reminder). |\n| `BID`         | string | Yes      | Block identifier associated with this content.                                     |\n| `user_id`     | string | Yes      | Unique identifier of the user posting the content.                                 |\n| `title`       | string | Optional | Title or short heading for the content.                                            |\n| `description` | string | Yes      | Main textual content to be stored and embedded.                                    |\n| `app_id`      | string | Optional | Identifier of the calling application.                                             |\n\n---\n\n## **Behavior**\n\n1. The API validates the user and floor context.\n2. Textual content (`title` and `description`) is ingested.\n3. Attached media files are processed and linked to the content.\n4. Embeddings are generated for:\n\n   * Text\n   * Supported media (where applicable)\n5. The content becomes part of the user’s **personal memory store**.\n6. The stored data is immediately available for querying and conversational retrieval.\n\n---\n\n## **Successful Response**\n\nOn success, the API confirms that:\n\n* The content has been stored\n* Embeddings have been generated\n* The memory item is available for future queries\n\nA success status and reference identifiers are returned.\n\n---\n\n## **Error Handling**\n\nThe API may return errors if:\n\n* Required fields are missing (`floor_id`, `user_id`, `description`)\n* Unsupported file formats are uploaded\n* The user does not have access to the specified floor\n* The request payload is malformed\n* Internal embedding or storage operations fail\n\n---\n\n## **Typical Use Cases**\n\n* Saving personal reminders\n* Posting notes or observations\n* Uploading documents for future reference\n* Creating a personal knowledge base\n* Feeding data into conversational agents\n\n---\n\n## **One-Line Summary**\n\n> Stores user-generated text and media into a personal POD, embeds it for semantic search, and makes it available for conversational querying.\n\n⚠️ Content Ingestion is Asynchronous\n\nThe Create Event API queues content for processing.\nA successful response indicates **acceptance**, not availability.\n\nNewly ingested content may take time to become searchable\nvia the Query API.\n\n",
        "operationId": "event",
        "tags": [
          "event"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "encoding": {
                "files": {
                  "contentType": "image/jpeg"
                }
              },
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "format": "binary",
                    "type": "string",
                    "description": "Attach relevant media here, which includes, jpg, mp3, pdf, mp4 files. More than one media can be selected",
                    "example": [
                      ""
                    ]
                  },
                  "input_info": {
                    "type": "string",
                    "description": "Input parameters, bid is optional",
                    "example": "{ \"floor_id\": \"my_floor\",   \"bid\": \"17845683456\",    \"user_id\": \"145623907625\",    \"title\": \"My floor\",    \"description\": \"My floor details\", \"app_id\": \"165434879028\"}"
                  }
                },
                "required": [
                  "input_info"
                ]
              },
              "examples": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "## **Successful Response (200 OK)**\n\nThis API follows an **asynchronous ingestion model**.\n\nOn success, the API **confirms acceptance of the content**, but **does not guarantee immediate availability** for retrieval or querying.\n\n### **What the response means**\n\nA `200 OK` response indicates that:\n\n* The request payload is valid\n* The content has been **accepted and queued** for processing\n* The content has been published to the internal event pipeline (RabbitMQ)\n* Embedding and storage will occur asynchronously\n\n⚠️ **Important:**\nThe newly posted content **will not appear immediately** in query or feed APIs.\n\n---\n\n## **Response Body (Example)**\n\n```json\n{\n  \"status\": \"success\",\n  \"message\": \"Content accepted for processing\"\n}\n```\n\n> No content data is returned in this response.\n\n---\n\n## **Availability & Retrieval Model**\n\nAfter receiving a successful response:\n\n1. The content is processed asynchronously\n2. The content becomes available in the floor feed **after processing completes**\n3. Developers must retrieve the content using:\n\n```\nGET /api/memory/recent/events\n```\n\n---\n\n## **Polling Requirement**\n\nBecause ingestion is asynchronous:\n\n* Developers **must poll** `/api/memory/recent/events`\n* Polling should be done using:\n\n  * `floor_id`\n  * timestamp or last-known event marker\n* Compare previously retrieved events with new responses to detect newly added content\n\nThis design ensures:\n\n* High ingestion throughput\n* Non-blocking uploads\n* Reliable embedding and storage pipelines\n\n---\n\n## **Typical Developer Flow**\n\n```text\nPOST /api/memory/events\n        ↓\n200 OK (content queued)\n        ↓\nPoll /api/memory/recent/events\n        ↓\nDetect new event\n        ↓\nUse content in /agent/memory/query\n```\n\n---\n\n## **Key Design Note (Why This Exists)**\n\nThis API is intentionally asynchronous to:\n\n* Support large files and multi-modal uploads\n* Avoid request timeouts during embedding\n* Enable scalable background processing\n* Keep write latency low\n\n---\n\n## **One-Line Summary**\n\n> Accepts user-generated text and media, queues it for asynchronous processing, and makes it available for retrieval via the recent events API after ingestion completes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                },
                "example": {
                  "success": "Sent to the queue"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Invalid block type or Invalid Bid type"
                        },
                        "path": {
                          "type": "string"
                        },
                        "timestamp": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "path",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                },
                "example": {
                  "error": {
                    "code": "XF_VALIDATION_ERROR",
                    "message": "Invalid BID",
                    "path": "/api/memory/events",
                    "timestamp": "2026-01-12T08:11:59.993079561Z"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "string"
          }
        },
        "required": [
          "success"
        ]
      },
...
    },
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "https://appfloor.in",
      "description": "Prod Env"
    }
  ],
  "security": []
}